# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ritme Salamat API",
        "description": "API Documentation for Ritme Salamat Application",
        "contact": {
            "email": "support@ritmesalamat.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server V1"
        }
    ],
    "paths": {
...
        "/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get user profile",
                "description": "Retrieve the authenticated user's profile information",
                "operationId": "7addf8d5759285de730066dc9e6ffce6",
                "responses": {
                    "200": {
                        "description": "Profile retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Sara",
                                                            "nullable": true
                                                        },
                                                        "mobile": {
                                                            "type": "string",
                                                            "example": "09123456789"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "profile": {
                                                    "properties": {
                                                        "birthday": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "1995-05-15"
                                                        },
                                                        "weight": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 60.5
                                                        },
                                                        "height": {
                                                            "type": "integer",
                                                            "example": 165
                                                        },
                                                        "period_duration": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "cycle_duration": {
                                                            "type": "integer",
                                                            "example": 28
                                                        },
                                                        "last_period_start": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "2024-12-01"
                                                        }
                                                    },
                                                    "type": "object",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Create or update user profile",
                "description": "Create or update the authenticated user's profile information including menstrual cycle data",
                "operationId": "df0f4e2a0885e2999baa496b7ce2f301",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "User's name",
                                        "type": "string",
                                        "example": "Sara"
                                    },
                                    "birthday": {
                                        "description": "Date of birth",
                                        "type": "string",
                                        "format": "date",
                                        "example": "1995-05-15"
                                    },
                                    "weight": {
                                        "description": "Weight in kg",
                                        "type": "number",
                                        "format": "float",
                                        "example": 60.5
                                    },
                                    "height": {
                                        "description": "Height in cm",
                                        "type": "integer",
                                        "example": 165
                                    },
                                    "period_duration": {
                                        "description": "How long does your period usually last? (days)",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "cycle_duration": {
                                        "description": "How long does your cycle usually last? (days)",
                                        "type": "integer",
                                        "example": 28
                                    },
                                    "last_period_start": {
                                        "description": "When did your last period start?",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Sara"
                                                        },
                                                        "mobile": {
                                                            "type": "string",
                                                            "example": "09123456789"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "profile": {
                                                    "properties": {
                                                        "birthday": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "1995-05-15"
                                                        },
                                                        "weight": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 60.5
                                                        },
                                                        "height": {
                                                            "type": "integer",
                                                            "example": 165
                                                        },
                                                        "period_duration": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "cycle_duration": {
                                                            "type": "integer",
                                                            "example": 28
                                                        },
                                                        "last_period_start": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "2024-12-01"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}